+++ examples/src/Server/Mythology/Character.hs
@@ -3,6 +3,7 @@
 {-# LANGUAGE DuplicateRecordFields #-}
 {-# LANGUAGE NamedFieldPuns #-}
 {-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE TypeOperators #-}
 
@@ -12,10 +13,13 @@ module Server.Mythology.Character
     Human (..),
     someHuman,
     someDeity,
+    Person,
   )
 where
 
-import Data.Morpheus.Types (GQLType (..))
+import Data.Morpheus.Kind (INTERFACE)
+import Data.Morpheus.Types (GQLType (..), interface)
+import Data.Proxy (Proxy (..))
 import Data.Text (Text)
 import GHC.Generics (Generic)
 import Server.Mythology.Place
@@ -23,13 +27,22 @@ import Server.Mythology.Place
     Realm (..),
   )
 
+newtype Person = Person {name :: Text}
+  deriving (Generic)
+
+instance GQLType Person where
+  type KIND Person = INTERFACE
+
 data Deity = Deity
   { name :: Text, -- Non-Nullable Field
     power :: Maybe Text, -- Nullable Field
     realm :: Realm,
     bornAt :: Maybe City
   }
-  deriving (Generic, GQLType)
+  deriving (Generic)
+
+instance GQLType Deity where
+  implements _ = [interface (Proxy @Person)]
 
 data Human m = Human
   { name :: m Text,